overpass-ai/requirements.txt
openai>=1.55.3
numpy==1.26.4
python-dotenv==1.0.0